The major challenges facing Quantitative Firm
In this study, the researcher examined the major challenges facing the setting up of Quantitative Finance Firm in an emerging economy like Nigeria. The paper reveals the challenges a new quantitative finance firm might likely face in Nigeria which are not limited to; deployment of ICT in operational services, corruption, human resources management, fraud prevention and monitoring system, liquidity management, full autonomy of the Central Bank of Nigeria, need for Capital Adequacy and technological innovation. It provides a relatively broad literature review of the challenges and empirically evaluated the influence of these challenges on the setting up of Quantitative Finance Firm in Nigeria.